test: move multi-precision to general tests + supersede sympy symbolic
