+++ pyproject.toml
@@ -23,16 +23,18 @@ dependencies = ["jax", "jaxlib", "jpu>=0.0.2", "equinox", "pint"]
 "Bug Tracker" = "https://github.com/exoplanet-dev/jaxoplanet/issues"
 
 [project.optional-dependencies]
-dev = ["nox", "ruff", "black", "pre-commit", "pytest", "sympy"]
-test = [
+dev = [
+    "nox",
+    "ruff",
+    "black",
+    "pre-commit",
     "pytest",
-    "pytest-xdist",
-    "exoplanet-core",
-    "batman-package",
+    "sympy",
     "mpmath",
     "gmpy2",
 ]
-test-math = ["sympy"]
+test = ["pytest", "pytest-xdist", "exoplanet-core", "batman-package"]
+test-math = ["mpmath", "gmpy2"]
 comparison = ["starry", "numpy<1.22", "tqdm", "xarray<2023.10.0"]
 docs = [
     "matplotlib",
